将decimal类型分数字段统一转换为double类型
- 修改Examina项目中的ExamScoreDetail、TrainingResultViewModel、ExamListViewModel等模型
- 修改ExaminaWebApplication项目中的ImportedQuestion、ImportedOperationPoint、ImportedSubject等实体类
- 修改BenchSuite项目中的ScoringModels和CSharpScoringService
- 更新所有API请求DTO中的分数字段类型
- 确保所有分数相关字段类型的完全统一

修改的文件：
- Examina/Models/ExamScoreDetail.cs
- Examina/ViewModels/Dialogs/ExamResultViewModel.cs
- Examina/ViewModels/TrainingResultViewModel.cs
- Examina/ViewModels/Pages/ExamListViewModel.cs
- Examina/Models/Api/SubmitExamScoreRequestDto.cs
- Examina/Models/Api/SubmitMockExamScoreRequestDto.cs
- ExaminaWebApplication/Models/ImportedExam/ImportedQuestion.cs
- ExaminaWebApplication/Models/ImportedExam/ImportedOperationPoint.cs
- ExaminaWebApplication/Models/ImportedExam/ImportedSubject.cs
- ExaminaWebApplication/Models/ImportedComprehensiveTraining/ImportedComprehensiveTrainingQuestion.cs
- ExaminaWebApplication/Models/ImportedSpecializedTraining/ImportedSpecializedTrainingOperationPoint.cs
- ExaminaWebApplication/Models/ImportedSpecializedTraining/ImportedSpecializedTrainingQuestion.cs
- ExaminaWebApplication/Models/Api/Student/SubmitExamScoreRequestDto.cs
- BenchSuite/Models/ScoringModels.cs
- BenchSuite/Services/CSharpScoringService.cs

解决问题：
- 完全统一所有项目中的分数字段数据类型
- 消除int、decimal、double混用导致的类型不一致问题
- 提高分数计算的精度和一致性
- 确保小数分数在所有层级都能正确处理

+++ ExamLab/Models/ImportExport/ExamExportDto.cs
@@ -113,7 +113,7 @@ public class ExamDto
     /// </summary>
     [JsonPropertyName("passingScore")]
     [XmlElement("PassingScore")]
-    public decimal PassingScore { get; set; } = 60.0m;
+    public double PassingScore { get; set; } = 60.0m;
 
     /// <summary>
     /// 是否随机题目顺序
@@ -285,7 +285,7 @@ public class SubjectDto
     /// 科目权重
     /// </summary>
     [JsonPropertyName("weight")]
-    public decimal Weight { get; set; } = 1.0m;
+    public double Weight { get; set; } = 1.0m;
 
     /// <summary>
     /// 科目配置
