+++ Examina/Models/ExamCompletion.cs
@@ -38,17 +38,17 @@ public class ExamCompletion
     /// <summary>
     /// 得分（可选）
     /// </summary>
-    public decimal? Score { get; set; }
+    public double? Score { get; set; }
 
     /// <summary>
     /// 最大得分（可选）
     /// </summary>
-    public decimal? MaxScore { get; set; }
+    public double? MaxScore { get; set; }
 
     /// <summary>
     /// 完成百分比（0-100）
     /// </summary>
-    public decimal? CompletionPercentage { get; set; }
+    public double? CompletionPercentage { get; set; }
 
     /// <summary>
     /// 用时（秒）
